+++ lib/actions/settings/shopSyrattachPluginSettingsSave.controller.php
@@ -16,16 +16,7 @@ class shopSyrattachPluginSettingsSaveController extends waJsonController
             throw new waException(_w('Access denied'));
         }
 
-        waLog::log('save', 'syrattach.log');
-
         $namespace = 'shop_syrattach';
-
-        $template = waRequest::post('syrattach_template', '', waRequest::TYPE_STRING);
-        if($template) {
-            $template_path = 'plugins/syrattach/templates/frontend_product.html';
-            $modified_template = waSystem::getInstance()->getDataPath($template_path, FALSE, 'shop', TRUE);
-            waFiles::write($modified_template, $template);
-        }
 
         $plugin = waSystem::getInstance()->getPlugin('syrattach');
         $settings = (array)$this->getRequest()->post($namespace);